Add common state to entity class
State includes position, whether the entity is shown near the screen
etc.

Includes refactor where closures can return values, and extends closure
use to memory-pools.

Finally incidentally adds the history_stack_flush method to purge the
stacks of all items.

+++ C/actors/ant.h
@@ -4,7 +4,7 @@
 
 typedef struct ant_struct ant;
 
-ant* ant_create(float x, float y);
+ant* ant_create(int x, int y);
 
 void ant_destroy(ant* ant);
